backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boxboxbox"
version = "0.1.0"
description = "AI-enhanced Formula 1 platform"
readme = "README.md"
authors = [
    {name = "Mantej Singh Arora", email = "mantej@example.com"},
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.3.0",
    "sqlalchemy>=2.0.20",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.7",
    "redis>=5.0.0",
    "celery>=5.3.4",
    "pandas>=2.1.0",
    "numpy>=1.25.2",
    "scikit-learn>=1.3.0",
    "torch>=2.0.1",
    "fastf1>=3.0.6",
    "requests>=2.31.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "fastapi-limiter>=0.1.5",
    "websockets>=11.0.3",
    "httpx>=0.24.1",
    "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "flake8>=6.1.0",
    "pre-commit>=3.3.3",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*" 